+++ modules/persistence/lib/schemas/troupe-invite-schema.js
@@ -12,6 +12,7 @@ module.exports = {
         type: { type: String, enum: ['email', 'github', 'twitter', 'gitlab'], required: true },
         emailAddress: { type: String, required: true },
         externalId: { type: String, required: true },
+        // This is used when the invite is accepted/rejected
         userId: { type: ObjectId, required: false },
         invitedByUserId: { type: ObjectId },
         secret: { type: String },
@@ -26,6 +27,7 @@ module.exports = {
     TroupeInviteSchema.index({ troupeId: 1 });
     TroupeInviteSchema.index({ type: 1, externalId: 1 });
     TroupeInviteSchema.index({ userId: 1 });
+    TroupeInviteSchema.index({ invitedByUserId: 1 });
     TroupeInviteSchema.index({ secret: 1 }, { unique: true, sparse: true });
     TroupeInviteSchema.index({ reminderSent: 1 }); // Prevent table scans on finding emails that need reminders sent out